+++ simfab.h
@@ -499,7 +499,7 @@ public:
 
 	sint32 liefere_an(const ware_besch_t *, sint32 menge);
 
-	void step(long delta_t);                  // fabrik muss auch arbeiten
+	void step(uint32 delta_t);                  // fabrik muss auch arbeiten
 	void neuer_monat();
 
 	char const* get_name() const;
