+++ simsys_opengl.cc
@@ -1138,7 +1138,7 @@ void ex_ord_update_mx_my()
 }
 
 
-unsigned long dr_time()
+uint32 dr_time()
 {
 	return SDL_GetTicks();
 }
